add Circle.js file, move cicle-function and import it
