+++ js-structure/js-structure_shapes/components/Circle/Circle.js
@@ -1,0 +1,8 @@
+export default function Circle() {
+    const circle = document.createElement("div");
+    circle.classList.add("circle");
+    circle.addEventListener("click", () => {
+      circle.style.backgroundColor = getRandomColor();
+    });
+    return circle;
+    };
